move docker stuff into class fixture to boost test run
